Compare chains of posterior trees

Performs a PERMANOVA (adonis2) to test for significant differences between posterior chains of transmission trees.

Usage

compare_chains(
  ...,
  f_tree = patristic,
  f_chain = euclidean,
  adonis2_args = list()
)

Arguments

...

Two or more posterior chains of transmission trees. Each chain is a list containing one or more data frames, where each data frame represents a transmission tree with columns from (infector) and to (infectee).

f_tree

A function to compute the pairwise distances between nodes within a tree. Must take a single tree (data frame) as input and return a square matrix where dimensions match the number of nodes. Default is patristic.

f_chain

A function to compute the distance between two chains of transmission trees. It must take two square matrices (produced by f_tree) as input and return a single numeric distance value. Default is euclidean.

adonis2_args

A list of additional arguments to pass to vegan::adonis2. Default is an empty list.

Value

A vegan::adonis2 object containing the results of the PERMANOVA.

Examples

set.seed(1)
#No difference in the chains
chainA <- replicate(10, igraph::as_long_data_frame(
  make_tree(n_cases = 10, R = 2, stochastic = TRUE)),
  simplify = FALSE)
chainB <- replicate(10, igraph::as_long_data_frame(
  make_tree(n_cases = 10, R = 2, stochastic = TRUE)),
  simplify = FALSE)
compare_chains(chainA, chainB)
#> Warning: number of items to replace is not a multiple of replacement length
#> Permutation test for adonis under reduced model
#> Permutation: free
#> Number of permutations: 999
#> 
#> (function (formula, data, permutations = 999, method = "bray", sqrt.dist = FALSE, add = FALSE, by = NULL, parallel = getOption("mc.cores"), na.action = na.fail, strata = NULL, ...) 
#>          Df SumOfSqs      R2      F Pr(>F)
#> Model     1    31.05 0.04356 0.8199  0.592
#> Residual 18   681.70 0.95644              
#> Total    19   712.75 1.00000              

#Difference in the chains
chainC <- replicate(10, igraph::as_long_data_frame(
  make_tree(n_cases = 10, R = 4, stochastic = TRUE)),
  simplify = FALSE)
compare_chains(chainA, chainB, chainC)
#> Warning: number of items to replace is not a multiple of replacement length
#> Permutation test for adonis under reduced model
#> Permutation: free
#> Number of permutations: 999
#> 
#> (function (formula, data, permutations = 999, method = "bray", sqrt.dist = FALSE, add = FALSE, by = NULL, parallel = getOption("mc.cores"), na.action = na.fail, strata = NULL, ...) 
#>          Df SumOfSqs      R2      F Pr(>F)    
#> Model     2   151.93 0.15283 2.4354  0.001 ***
#> Residual 27   842.20 0.84717                  
#> Total    29   994.13 1.00000                  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1